requirements.txt
numpy==1.19.5
tensorflow==2.6.0
streamlit==1.0.0
matplotlib==3.4.2
Pillow==8.3.2
